Rename closures -> coordinate-transformation
