+++ exercises/concept/coordinate-transformation/.meta/exemplar.js
