+++ exercises/concept/coordinate-transformation/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@exercism/javascript-closures",
+  "name": "@exercism/javascript-coordinate-transformation",
   "description": "Exercism concept exercise on closures",
   "author": "Tim Austin",
   "private": true,
@@ -7,7 +7,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/exercism/javascript",
-    "directory": "exercises/concept/closures"
+    "directory": "exercises/concept/coordinate-transformation"
   },
   "devDependencies": {
     "@babel/cli": "^7.14.5",
